requirements.txt
alembic==1.14.0
anyio==4.7.0
APScheduler==3.6.3
attrs==25.1.0
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.13.0
cachetools==4.2.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.6
MarkupSafe==3.0.2
msgspec==0.18.6
numpy==2.2.2
outcome==1.3.0.post0
pandas==2.2.3
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.28.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
tornado==6.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.0
wsproto==1.2.0
